rename histogram.* to histogram_impl.* for consistency.
