+++ source/common/stats/isolated_store_impl.cc
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "common/common/utility.h"
-#include "common/stats/histogram.h"
+#include "common/stats/histogram_impl.h"
 #include "common/stats/utility.h"
 
 namespace Envoy {
